refactor: reorganize integration structure by moving dataSource and DGS schemas to a new integration directory
