+++ packages/components/src/interfaces/integration/dataSource.ts
@@ -1,0 +1,4 @@
+export const DATA_SOURCE_TYPE = {
+  native: "native",
+  dgs: "dgs",
+} as const
