+++ packages/components/src/interfaces/integration/index.ts
@@ -1,0 +1,2 @@
+export * from "./dataSource"
+export * from "./dgs"
